#551 create new library for lobby sql tools
